+++ src/main/java/de/codescape/jira/plugins/scrumpoker/model/ProjectActivation.java
@@ -1,0 +1,20 @@
+package de.codescape.jira.plugins.scrumpoker.model;
+
+public enum ProjectActivation {
+
+    /**
+     * Inherit activation for the current project from the global settings.
+     */
+    INHERIT,
+
+    /**
+     * Scrum Poker is explicitly activated for the current project regardless of the global settings.
+     */
+    ACTIVATE,
+
+    /**
+     * Scrum Poker is explicitly disabled for the current project regardless of the global settings.
+     */
+    DISABLE;
+
+}